enhance: `data/table`: update `Table.RowCellCounts()` to strict `map[uint]uint` return value

+++ data/table/row.go
@@ -6,10 +6,10 @@ import (
 
 // RowCellCounts returns a `map[int]int` where the key is the cell count
 // and the value is the number of rows.
-func (tbl *Table) RowCellCounts() map[int]int {
-	mii := map[int]int{}
+func (tbl *Table) RowCellCounts() map[uint]uint {
+	mii := map[uint]uint{}
 	for _, row := range tbl.Rows {
-		mii[len(row)] += 1
+		mii[uint(len(row))] += 1
 	}
 	return mii
 }
